final
final

+++ src/BubleSortConsistently.java
@@ -1,13 +1,13 @@
 // Последовательный алгоритм чет-нечетной сортировки
 public class BubleSortConsistently {
 
-    public static void bubbleSort(int[] A, int left, int right) {
+    public static void bubbleSort(long[] A, int left, int right) {
         boolean isSorted = false;
         while (!isSorted) {
             isSorted = true;
             for (int i = 0; i <= A.length - 2; i += 2) {
                 if (A[i] > A[i + 1]) {
-                    int temp = A[i];
+                    long temp = A[i];
                     A[i] = A[i + 1];
                     A[i + 1] = temp;
                     isSorted = false;
@@ -15,7 +15,7 @@ public class BubleSortConsistently {
             }
             for (int i = 1; i <= A.length - 2; i += 2) {
                 if (A[i] > A[i + 1]) {
-                    int temp = A[i];
+                    long temp = A[i];
                     A[i] = A[i + 1];
                     A[i + 1] = temp;
                     isSorted = false;
